Auto-rebuild history completion on schema change.
When changing schema and bumping USER_VERSION, the completion would be
rebuilt but the schema would not be updated. Instead of using
delete_all, which removes content but leaves the table in place, we need
to use drop to remove the table itself, so the new table can be created.

This also adds more logging around history regeneration.